+++ cmd/nodecmd/node.go
@@ -36,8 +36,6 @@ rest of the commands to maintain your node and make your node a Subnet Validator
 	cmd.AddCommand(NewValidateCmd())
 	// node sync cluster --subnet subnetName
 	cmd.AddCommand(newSyncCmd())
-	// node stop
-	cmd.AddCommand(newStopCmd())
 	// node terminate
 	cmd.AddCommand(newTerminateCmd())
 	// node status cluster
